Lift out poa verification
