+++ modules/claims_api/app/controllers/claims_api/application_controller.rb
@@ -4,7 +4,6 @@ module ClaimsApi
   class ApplicationController < ::OpenidApplicationController
     skip_before_action :set_tags_and_extra_context, raise: false
     before_action :log_request
-    before_action :verify_power_of_attorney, if: :poa_request?
     before_action :verify_mvi
 
     private
@@ -60,22 +59,11 @@ module ClaimsApi
       end
     end
 
-    def verify_power_of_attorney
-      verifier = EVSS::PowerOfAttorneyVerifier.new(target_veteran)
-      verifier.verify(@current_user)
-    end
-
     def verify_mvi
       unless target_veteran.mvi_record?
         render json: { errors: [{ detail: 'Not found' }] },
                status: :not_found
       end
-    end
-
-    def poa_request?
-      # if any of the required headers are present we should attempt to use headers
-      headers_to_check = ['HTTP_X_VA_SSN', 'HTTP_X_VA_Consumer-Username', 'HTTP_X_VA_Birth_Date']
-      (request.headers.to_h.keys & headers_to_check).length.positive?
     end
   end
 end
