+++ modules/claims_api/app/controllers/claims_api/v1/forms/intent_to_file_controller.rb
@@ -1,11 +1,14 @@
 # frozen_string_literal: true
 
 require_dependency 'claims_api/intent_to_file_serializer'
+require_dependency 'claims_api/concerns/poa_verfication'
 
 module ClaimsApi
   module V1
     module Forms
       class IntentToFileController < BaseFormController
+        include ClaimsApi::PoaVerfication
+
         before_action { permit_scopes %w[claim.write] }
         before_action :check_future_type, only: [:submit_form_0966]
         skip_before_action :validate_json_schema, only: [:active]
